loan_tracker/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.6
django-cors-headers==4.4.0
django-filter==24.2
django-grappelli==4.0.1
django-render-block==0.9.2
django-rest-swagger==2.2.0
django-templated-email==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.1
pillow==10.3.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
reportlab==4.2.0
requests==2.32.3
setuptools==70.1.0
simplejson==3.19.2
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2